refactor out deaths per capita chart
